+++ cell_functions.h
@@ -7,7 +7,7 @@ extern void (**CELL_UPDATE_FUNCTIONS)(Generation *gen, Cell *cell);
 
 extern void (**CELL_INIT_FUNCTIONS)(Generation *gen, Cell *cell);
 
-extern void add_particle(size_t particle_index,
+extern void add_particle(int particle_index,
                          void (*init_func)(Generation *gen, Cell *cell),
                          void (*update_func)(Generation *gen, Cell *cell));
